+++ src/components/ui/Modal.svelte
@@ -21,7 +21,7 @@
             <div class="bg-gray-600 pt-3 pb-6">
               {#if title}
                 <div id="modal-title" class="flex flex-row gap-3 items-center pl-4 pr-2 pb-3">
-                  <div class="grow min-w-0 ellipsis text-gray-400 select-none">{title}</div>
+                  <div class="grow min-w-0 truncate text-gray-400 select-none">{title}</div>
                   <Button
                     rounded
                     icon={biX}